Extract client method for user account in AccountClient class
