+++ tests/clients/BarracksClient.spec.js
@@ -32,6 +32,25 @@ describe('Barracks', () => {
     barracks.v2Enabled = false;
   });
 
+  describe('#constructor()', () => {
+
+    let barracksClient;
+
+    beforeEach(() => {
+      barracksClient = undefined;
+    });
+
+    it('should initialize methods from accountClient when constructor called', () => {
+      // When
+      const barracksClient = new Barracks();
+
+      // Then
+      expect(barracksClient).to.have.property('authenticate').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getAccount').and.to.be.a('function');
+      expect(barracksClient).to.have.property('setGoogleAnalyticsTrackingId').and.to.be.a('function');
+    });
+  });
+
   describe('#editUpdate()', () => {
 
     const originalUpdate = {
@@ -207,139 +226,6 @@ describe('Barracks', () => {
       });
     });
 
-  });
-
-  describe('#authenticate()', () => {
-
-    it('should return an error message when authentication fails', done => {
-      // Given
-      const username = 'user';
-      const password = 'password';
-      const error = { message: 'Login failed' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.authenticate(username, password).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('login', {
-          body: { username, password }
-        });
-        done();
-      });
-    });
-
-    it('should return a token when authentication succeed', done => {
-      // Given
-      const username = 'user';
-      const password = 'password';
-      const response = { headers: { 'x-auth-token': token } };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.authenticate(username, password).then(result => {
-        expect(result).to.be.equals(token);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('login', {
-          body: { username, password }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#getAccount()', () => {
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.getAccount(token).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('me', {
-          headers: { 'x-auth-token': token }
-        });
-        done();
-      });
-    });
-
-    it('should return a token when authentication succeed', done => {
-      // Given
-      const account = { apiKey: 'qwertyuiop', username: 'coucou' };
-      const response = { body: account };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getAccount(token).then(result => {
-        expect(result).to.be.equals(account);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('me', {
-          headers: { 'x-auth-token': token }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#setGoogleAnalyticsId()', () => {
-
-    const analyticsId = 'UA-12345678-1';
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.setGoogleAnalyticsTrackingId(token, analyticsId).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly(
-          'setGoogleAnalyticsTrackingId',
-          {
-            headers: { 'x-auth-token': token },
-            body: { value: analyticsId }
-          }
-        );
-        done();
-      });
-    });
-
-    it('should return the server response when request success', done => {
-      // Given
-      const account = { apiKey: 'qwertyuiop', username: 'coucou', gaTrackingId: analyticsId };
-      const response = { body: account };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.setGoogleAnalyticsTrackingId(token, analyticsId).then(result => {
-        expect(result).to.be.equals(account);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly(
-          'setGoogleAnalyticsTrackingId',
-          {
-            headers: { 'x-auth-token': token },
-            body: { value: analyticsId }
-          }
-        );
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
   });
 
   describe('#getSegment()', () => {
